requirements.txt
Flask
python-binance
pandas
numpy
joblib
tensorflow
ta
scikit-learn
